mpegmap/TC_5.1.4-10_RegionRatingTableDataCues.html

<!doctype html>
<html>
  <head>
    <title>TC 5.1.4-10 Region Rating Table DataCues</title>
    <link rel="author" title="CableLabs" href="html5@cablelabs.com">
    <meta name="flags" content="[requirement flags]">
    <meta name="assert" content="MPEG-2 SPTS, Region Rating Table - UA creates DataCues, attributes set.">
    <meta name="timeout" content="long">
    <script src="testmedia.js"></script>
    <!-- Source HTTP server local copy of World Wide Web Consortium (W3C) Test Harness. --> 
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <!-- Script Version: BETA -->

    <!-- About the Media Resource
           To verify DOM elements the details of the test media must be known prior to test execution.
           This test script is bound to the test media file sourced in the script.
           Sourcing a differnt test media file, or modifying that file, may invalidate the test.

           The media is an MPEG-2, Single Program Transport Stream
           containing two EISS streams, one program insertion cue stream.
    -->
  </head>


  <body>
    <p><a href="http://html5.cablelabs.com">spec reference</a></p>
    <p><a href="http://html5.cablelabs.com/dlna-rui-test/dlna">test assertion</a></p>
    <video id="v" width="400" height="300" autoplay controls></video>
    <div id="log"></div>


    <script>
      var test = async_test("Region Rating Table DataCue attributes.");
      var vid = document.getElementById("v");
      vid.src = mediaServerURL() + "ClosedCaption.ts"; 

      var index = -1;
      var pid = '8187';  // Region Rating Table PID

      // TextTrack[index].cues is a list.  Each cue in the list will have a unique data attribute value.
      // This script tests that all cues are instanceof ArrayBuffer.
      // This script tests TextTrack[index].cues[0].data only for value.
      var data = [199, 240, 135, 0, 0, 203, 0, 0, 0, 0, 11, 0, 0, 255, 251, 233, 0, 0, 0, 88, 240, 0, 0, 4, 241, 255, 233, 0, 0, 0, 50, 240, 0, 3, 1, 255, 251, 224, 0, 0, 3, 133, 240, 0, 1, 0, 240, 0, 233, 0, 0, 0, 90, 240, 0, 1, 1, 240, 1, 233, 0, 0, 0, 90, 240, 0, 1, 2, 240, 2, 233, 0, 0, 0, 166, 240, 0, 1, 3, 240, 3, 233, 0, 0, 0, 90, 240, 0, 2, 0, 242, 0, 233, 0, 0, 0, 57, 240, 0, 2, 1, 242, 1, 233, 0, 0, 0, 57, 240, 0, 2, 2, 242, 2, 233, 0, 0, 0, 114, 240, 0, 2, 3, 242, 3, 233, 0, 0, 0, 57, 240, 0, 240, 0, 58, 89, 97, 242];


      vid.addEventListener("playing", function() {
        vid.pause();

        setTimeout(function() {
          for (i=0; i< vid.textTracks.length; i++) {
            if (vid.textTracks[i].id == pid) {
              vid.textTracks[i].mode = 'hidden';
              index = i;
              break;
            }
          }
        }, 6000);

        setTimeout(function() {
          if (index == -1) {
            test.step(function() {
              assert_unreached("Did not find Region Rating Table TextTrack.id (PID) " + pid +  " in DOM.");
            }, "Did not find Region Rating Table TextTrack.id (PID) " + pid +  " in DOM.");
            test.done();

          } else {
            test.step(function() {
              for (i=0; i< vid.textTracks[index].cues.length; i++) {

                var startTime = vid.textTracks[index].cues[i].startTime;
                assert_equals(startTime, 0, "DataCues.startTime");
                var endTime = vid.textTracks[index].cues[i].endTime;
                assert_greater_than_equal(endTime, startTime, "DataCues.endTime");

                assert_true(vid.textTracks[index].cues[i].data instanceof ArrayBuffer, "DataCues[i] is ArrayBuffer");

                if (i==0) {
                  var cueData = new Uint8Array(vid.textTracks[index].cues[0].data);
                  assert_array_equals(data, cueData, "DataCues[0] data");
                }

                assert_equals(vid.textTracks[index].cues[i].pauseOnExit, false, "DataCues.pauseOnExit");
                assert_equals(vid.textTracks[index].cues[i].text, null, "DataCues.text");
              }
            }, "Region Rating Table DataCues attributes.");

            test.done();
          }
        }, 12000);
      }, false);

    </script>
  </body>
</html>
